+++ src/engine/util/util_draw.h
@@ -16,7 +16,7 @@ namespace Util
              SDL_Color color);
 
     void
-    DrawGrid(SDL_Renderer *renderer, Uint32 spacing_x, Uint32 spacing_y, float start_x, float start_y, Uint32 n_rows, Uint32 n_cols,
+    DrawGrid(SDL_Renderer *renderer, float spacing_x, float spacing_y, float start_x, float start_y, Uint32 n_rows, Uint32 n_cols,
              SDL_Color color);
 
     void
